begining rework
