--- src/main/kotlin/io/github/realyusufismail/armourandtoolsmod/blocks/fusion/OutputSlotItemHandler.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2023 RealYusufIsmail.
- *
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- *
- * you may not use this file except in compliance with the License.
- *
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */ 
-package io.github.realyusufismail.armourandtoolsmod.blocks.fusion
-
-import net.minecraft.world.item.ItemStack
-import net.minecraftforge.items.IItemHandler
-import net.minecraftforge.items.SlotItemHandler
-
-class OutputSlotItemHandler(itemHandler: IItemHandler) : SlotItemHandler(itemHandler, 3, 138, 40) {
-
-    override fun mayPlace(stack: ItemStack): Boolean {
-        return false
-    }
-}
